+++ scripts/format.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+SOURCES="piccolo tests"
+
+echo "Running isort..."
+isort $SOURCES
+echo "-----"
+
+echo "Running black..."
+black $SOURCES